requirements.txt
# dispatcher
twisted
apscheduler

# crawl_daemon
scrapy
python-memcached

# web

# all
pika
flask-mongoengine

# optional
netifaces
